scripts_py/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.3.2
pandas==2.2.2
fake-useragent==1.5.1
opencv-python==4.9.0.80
numpy==1.26.4
pillow==11.2.1
shapely==2.1.1 --prefer-binary
geojson==3.1.0
tqdm==4.66.4

